Swasya AI backend/llama/requirements.txt
llama-index
llama-index-llms-openai
pymongo
python-dotenv
